Utilities cleanup
